requirements.txt
# =============================================================================
# SPECTER — Dependencias Python
# =============================================================================

# --- Fase 1: Infraestrutura & Ingestao ---
celery[redis]>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
httpx>=0.27.0
tenacity>=8.2.0
python-dotenv>=1.0.0
structlog>=24.1.0
loguru>=0.7.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# --- Fase 2: Feature Engineering ---
rapidfuzz>=3.6.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# --- Fase 3: ML ---
xgboost>=2.0.0
scikit-learn>=1.4.0
optuna>=3.5.0
imbalanced-learn>=0.12.0
pandas>=2.2.0
pyarrow>=15.0.0
shap>=0.45.0
joblib>=1.3.0

# --- Fase 3.3: LLM ---
anthropic>=0.18.0
openai>=1.12.0
pydantic[email]>=2.6.0

# --- Fase 4: API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# --- Testes ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
